+++ api/libs/recipe/src/dto/recipe/recipe-response.dto.ts
@@ -2,7 +2,7 @@ import {
   CreatedResponse,
   OkResponse,
 } from '@app/common/dto/success-response.dto';
-import { Recipe } from '../entities/recipe.entity';
+import { Recipe } from '../../entities/recipe.entity';
 import {
   RecipeDto,
   RecipeListViewResponseDto,
