+++ api/libs/recipe/src/repositories/recipe-bookmark.repository.ts
@@ -3,11 +3,11 @@ import {
   RecipeBookmark,
   RecipeBookmarkDocument,
 } from '../entities/recipe-bookmark.entity';
-import { CreateRecipeBookmarkDto } from '../dto/modify-recipe-bookmark.dto';
+import { CreateRecipeBookmarkDto } from '../dto/recipe-bookmark/modify-recipe-bookmark.dto';
 import {
   FilterRecipeBookmarkDto,
   RecipeBookmarksAndCountDto,
-} from '../dto/filter-recipe-bookmark.dto';
+} from '../dto/recipe-bookmark/filter-recipe-bookmark.dto';
 import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
 import { transferObjectId } from '@app/common/utils/transfer-objectId';
